Fix crash on FingerprintFindActivity because of .so files
The fingerprint SDK we use relies on several .so (object files, compiled from C/C++ source code) files. But whenever we opened the `FingerprintFindActivity`, which uses these files, the app would crash. We discovered that this was because Gradle couldn't find the .so files. So I put these files in a more convenient location and told Gradle where to find them, and everything works now.